pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plex-letterboxd"
version = "0.6"
description = "Export Plex watch history to Letterboxd-compatible CSV format"
readme = "README.md"
license = "MIT"
authors = [
  { name = "Wyatt Brege", email = "wyatt@brege.org" },
]
requires-python = ">=3.10"
dependencies = [
  "plexapi>=4.17.1",
  "PyYAML>=6.0.2",
  "click>=8.1.7",
  "pydantic>=2.0",
]

[project.urls]
Repository = "https://github.com/brege/plex-letterboxd"

[project.scripts]
plex-letterboxd = "plex_letterboxd.cli:main"

[dependency-groups]
dev = [
  "ty>=0.0.18",
]

[tool.ty.src]
include = ["plex_letterboxd"]
respect-ignore-files = true

[tool.hatch.build.targets.wheel]
packages = ["plex_letterboxd"]

[tool.hatch.build.targets.sdist]
include = ["/plex_letterboxd", "/config.example.yaml", "/systemd"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E203"]

[tool.ruff.lint.isort]
split-on-trailing-comma = true
